+++ packages/gamut/src/GridForm/GridFormInputGroup/__fixtures__/renderers.tsx
@@ -1,8 +1,8 @@
-import { GamutThemeProvider } from '@codecademy/gamut-styles';
+import { theme } from '@codecademy/gamut-styles';
+import { ThemeProvider } from '@emotion/react';
 import { mount } from 'enzyme';
 import React from 'react';
 
-import { ToolTipProps } from '../../../ToolTip';
 import {
   stubCheckboxField,
   stubFileField,
@@ -25,10 +25,9 @@ import { GridFormRadioGroupInput } from '../GridFormRadioGroupInput';
 import { GridFormSelectInput } from '../GridFormSelectInput';
 import { GridFormTextArea } from '../GridFormTextArea';
 import { GridFormTextInput } from '../GridFormTextInput';
-import { GridFormToolTip } from '../GridFromToolTip';
 
 const mountWithTheme = (component: React.ReactNode) => {
-  return mount(<GamutThemeProvider>{component}</GamutThemeProvider>);
+  return mount(<ThemeProvider theme={theme}>{component}</ThemeProvider>);
 };
 
 export const renderGridFormSelectInput = (
@@ -104,10 +103,6 @@ export const renderGridFormCheckboxInput = (
   );
 };
 
-export const renderGridFormToolTip = (extraProps: ToolTipProps) => {
-  return mountWithTheme(<GridFormToolTip {...extraProps} />);
-};
-
 export const getComponent = (componentName: string, extraProps: any) => {
   switch (componentName) {
     case 'GridFormTextInput':
@@ -122,8 +117,6 @@ export const getComponent = (componentName: string, extraProps: any) => {
       return renderGridFormFileInput(extraProps);
     case 'GridFormCheckboxInput':
       return renderGridFormCheckboxInput(extraProps);
-    case 'GridFormToolTip':
-      return renderGridFormToolTip(extraProps);
     default:
       throw new Error(`Unknown component name: ${componentName}`);
   }