+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.2.33](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.32...@codecademy/gamut-kit@0.2.33) (2021-03-24)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.2.32](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.31...@codecademy/gamut-kit@0.2.32) (2021-03-23)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.2.31](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.30...@codecademy/gamut-kit@0.2.31) (2021-03-23)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.2.30](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.29...@codecademy/gamut-kit@0.2.30) (2021-03-23)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.2.29](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.28...@codecademy/gamut-kit@0.2.29) (2021-03-23)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
 ### [0.2.28](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.2.27...@codecademy/gamut-kit@0.2.28) (2021-03-22)
 
 **Note:** Version bump only for package @codecademy/gamut-kit