+++ packages/gamut-labs/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [13.1.3](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@13.1.2...@codecademy/gamut-labs@13.1.3) (2021-03-24)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
+### [13.1.2](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@13.1.1...@codecademy/gamut-labs@13.1.2) (2021-03-23)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
+### [13.1.1](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@13.1.0...@codecademy/gamut-labs@13.1.1) (2021-03-23)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
+## [13.1.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@13.0.5...@codecademy/gamut-labs@13.1.0) (2021-03-23)
+
+
+### Features
+
+* **GamutProvider:** Adds single provider for Theme + Cache + Globals + Assets [WEB-1352] [GM-173] ([9eac700](https://github.com/Codecademy/client-modules/commit/9eac7009b4d773621a0993cc125d6878b76977ce))
+
+
+
+### [13.0.5](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@13.0.4...@codecademy/gamut-labs@13.0.5) (2021-03-23)
+
+**Note:** Version bump only for package @codecademy/gamut-labs
+
+
+
+
+
 ### [13.0.4](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@13.0.3...@codecademy/gamut-labs@13.0.4) (2021-03-22)
 
 **Note:** Version bump only for package @codecademy/gamut-labs